+++ lib/R7RS.scm
@@ -108,8 +108,6 @@
  (array->list (vector.slice start end)))
 
 ;; -----------------------------------------------------------------------------
-
-;; -----------------------------------------------------------------------------
 (%range-function
  (string->vector string)
  "(string->list string)
@@ -138,6 +136,22 @@
        (slice start end)
        (map (lambda (char.valueOf)))
        (join "")))
+
+;; -----------------------------------------------------------------------------
+(%range-function
+ (vector-fill! vector fill)
+ "(vector-fill! vector fill)
+  (vector-fill! vector fill start)
+  (vector-fill! vector fill start end)
+
+  Fill vector with a given value in given range. If start is not given is start
+  at 0. If end is not given it fill till the end if the vector."
+ (typecheck "vector->fill!" vector "array")
+ (let recur ((n (- end start)))
+    (if (>= n start)
+        (begin
+          (set-obj! vector n fill)
+          (recur (- n 1))))))
 
 ;; -----------------------------------------------------------------------------
 (define-syntax let*-values